% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{convertLegacyCoseq}
\alias{convertLegacyCoseq}
\title{Convert legacy coseq objects}
\usage{
convertLegacyCoseq(object, digits = 3)
}
\arguments{
\item{object}{Object of S3 class \code{coseq} arising from a call to previous versions of
coseq (< 0.99.1)}

\item{digits}{integer indicating the number of decimal places (round) to retain in results.}
}
\value{
Converted object of S4 class \code{coseqResults} compatible with
recent versions of coseq (>= 0.99.1)
}
\description{
Convert legacy coseq S3 class objects to coseqResults S4 class objects
}

