% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coseqUsersGuide.R
\name{coseqUsersGuide}
\alias{coseqUsersGuide}
\title{View coseq User's Guide}
\usage{
coseqUsersGuide(view = TRUE)
}
\arguments{
\item{view}{logical, should the document be opened using the default PDF
document reader?}
}
\value{
Character string giving the file location. If \code{view=TRUE}, the
PDF document reader is started and the User's Guide is opened, as a side
effect.
}
\description{
Finds the location of the coseq User's Guide and optionally opens it.
}
\details{
The function \code{vignette("coseq")} will find the short coseq
Vignette which describes how to obtain the coseq User's Guide. The
User's Guide is not itself a true vignette because it is not automatically
generated using \code{\link{Sweave}} during the package build process. This
means that it cannot be found using \code{vignette}, hence the need for this
special function.

If the operating system is not Windows, the PDF viewer used is
given by \code{Sys.getenv("R_PDFVIEWER")}. The PDF viewer can be
changed using \code{Sys.putenv(R_PDFVIEWER=)}.

Note that this function was adapted from that defined by Gordon Smyth in the
edgeR package.
}
\examples{

# To get the location:
coseqUsersGuide(view=FALSE)
# To open in pdf viewer:
\dontrun{coseqUsersGuide()}

}
\author{
Gordon Smyth
}
\seealso{
\code{\link{system}}
}
\keyword{documentation}

