% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coseq.R
\name{plot.coseq}
\alias{plot.coseq}
\title{Visualize results from coseq clustering}
\usage{
\method{plot}{coseq}(x, y_profiles = NULL, K = NULL, threshold = 0.8,
  conds = NULL, average_over_conds = FALSE, graphs = c("logLike", "ICL",
  "profiles", "boxplots", "probapost_boxplots", "probapost_barplots",
  "probapost_histogram"), order = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"coseq"}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables to be used for graphing results (optional for
\code{logLike}, \code{ICL}, \code{probapost_boxplots}, and \code{probapost_barplots},
and by default takes value \code{x$tcounts} if \code{NULL})}

\item{K}{If desired, the specific model to use for plotting. If \code{NULL},
the model chosen by ICL will be plotted}

\item{threshold}{Threshold used for maximum conditional probability; only observations
with maximum conditional probability greater than this threshold are visualized}

\item{conds}{Condition labels, if desired}

\item{average_over_conds}{If \code{TRUE}, average values of \code{y_profiles} within
each condition identified by \code{conds} for the \code{profiles} and \code{boxplots}
plots}

\item{graphs}{Graphs to be produced, one (or more) of the following: 
\code{"logLike"} (log-likelihood plotted versus number of clusters),
\code{"ICL"} (ICL plotted versus number of clusters), 
\code{"profiles"} (line plots of profiles in each cluster), \code{"boxplots"} 
(boxplots of profiles in each cluster), \code{"probapost_boxplots"} (boxplots of
maximum conditional probabilities per cluster), \code{"probapost_barplots"} 
(number of observations with a maximum conditional probability greater than 
\code{threshold} per cluster), \code{"probapost_histogram"} (histogram of maximum
conditional probabilities over all clusters) ...}

\item{order}{If \code{TRUE}, order clusters in \code{probapost_boxplot} by median and
\code{probapost_barplot} by number of observations with maximum conditional probability
greater than \code{threshold}}

\item{...}{Additional optional plotting arguments}
}
\description{
Plot a coseq object.
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau
}

