% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NormMixClus_K.R
\name{plot.NormMixClus_K}
\alias{plot.NormMixClus_K}
\title{Plot NormMixClus_K object}
\usage{
\method{plot}{NormMixClus_K}(x, y_profiles, K = NULL, threshold = 0.8,
  conds = NULL, average_over_conds = TRUE, graphs = c("profiles",
  "boxplots"), ...)
}
\arguments{
\item{x}{An object of class \code{"NormMixClus_K"}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables to be used for graphing}

\item{K}{If desired, the specific cluster number to use for plotting. If \code{NULL},
all clusters will be visualized}

\item{threshold}{Threshold used for maximum conditional probability; only observations
with maximum conditional probability greater than this threshold are visualized}

\item{conds}{Condition labels, if desired}

\item{average_over_conds}{If \code{TRUE}, average values of \code{y_profiles} within
each condition identified by \code{conds}}

\item{graphs}{Graphs to be produced, one (or more) of the following: 
\code{"profiles_K"}, ...}

\item{...}{Additional optional plotting arguments}
}
\description{
Blah blah blah
}

