% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NormMixClus.R
\name{summary.NormMixClus}
\alias{summary.NormMixClus}
\title{Summarize results from clustering using a Normal mixture model}
\usage{
\method{summary}{NormMixClus}(object, y_profiles, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{"NormMixClus"}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{digits}{Integer indicating the number of decimal places to be used
for mixture model parameters}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Normal
mixture model estimated using \code{NormMixClus}.
}
\details{
The summary function for an object of class \code{"NormMixClus"}
provides the number of clusters selected for the ICL
model selection approach.
}
\examples{
## Simulate toy data, n = 300 observations
set.seed(12345)
countmat <- matrix(runif(300*4, min=0, max=500), nrow=300, ncol=4)
countmat <- countmat[which(rowSums(countmat) > 0),]
profiles <- transform_RNAseq(countmat, norm="none", 
                             transformation="arcsin")$tcounts

conds <- rep(c("A","B","C","D"), each=2)

## Run the Normal mixture model for K = 2,3
run <- NormMixClus(y=profiles, K=2:3, iter=5)

## Run the Normal mixture model for K=2
run2 <- NormMixClus_K(y=profiles, K=2, iter=5)

## Re-estimate mixture parameters for the model with K=2 clusters
param <- NormMixParam(run2, y_profiles=profiles)

## Summary of results
summary(run, y_profiles=profiles)
}
\author{
Andrea Rau
}
\references{
Rau, A. and Maugis-Rabusseau, C. (2016) Transformation and model choice for
co-expression analayis of RNA-seq data. bioRxiv, doi: http://dx.doi.org/10.1101/065607.
}
\seealso{
\code{\link{NormMixClus}}, \code{\link{NormMixClus_K}}
}
\keyword{methods}

