% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PoisMixClus.R
\name{summary.PoisMixClus}
\alias{summary.PoisMixClus}
\title{Summarize results from clustering using a Poisson mixture model}
\usage{
\method{summary}{PoisMixClus}(object, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{"PoisMixClus"}}

\item{digits}{Integer indicating the number of decimal places to be used
for mixture model parameters}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Poisson
mixture model estimated using \code{coseq} (which indirectly calls
\code{HTSCluster}.
}
\details{
The summary function for an object of class \code{"PoisMixClus"}
provides the number of clusters selected according to the user-defined
model selection criterion.
}
\examples{
## Simulate toy data, n = 300 observations
set.seed(12345)
countmat <- matrix(round(runif(300*4, min=0, max=500)), nrow=300, ncol=4)
countmat <- countmat[which(rowSums(countmat) > 0),]
conds <- rep(c("A","B","C","D"), each=2)

## Run the Poisson mixture model for K = 2,3
run <- coseq(y=countmat, K=2:3, iter=5, model="Poisson")
plot(run)
summary(run)
}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{coseq}}
}
\keyword{methods}

