% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareARI.R
\name{compareARI}
\alias{compareARI}
\title{Title}
\usage{
compareARI(x, parallel = FALSE, BPPARAM = bpparam(), plot = TRUE, ...)
}
\arguments{
\item{x}{blah}

\item{parallel}{blah}

\item{BPPARAM}{blah}

\item{plot}{blah}

\item{...}{Additional parameters for corrplot}
}
\value{
Matrix of adjusted rand index values calculated between each pair of models.
}
\description{
Title
}

