% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareICL.R
\name{compareICL}
\alias{compareICL}
\title{Compare corrected ICL values after data transformation}
\usage{
compareICL(x)
}
\arguments{
\item{x}{A list made up of \code{coseq} objects. At the current time, this function
only supports the comparison of \code{coseq} objects using \code{model="Normal"} and
\code{transformation = c("arcsin", "logit", "logMedianRef")}}
}
\value{
A plot of corrected ICL values for the models included in \code{x} (the list
of \code{coseq} objects)
}
\description{
Compare the corrected ICL values after applying the arcsin, logit, and logMedianRef
transformations in a coseq analysis
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau
}

