% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareICL.R
\name{compareICL}
\alias{compareICL}
\title{Compare corrected ICL values after data transformation}
\usage{
compareICL(x)
}
\arguments{
\item{x}{A list made up of \code{coseq} objects. At the current time, this function
only supports the comparison of \code{coseq} objects using \code{model="Normal"} and
\code{transformation = c("arcsin", "logit", "logMedianRef")}}
}
\value{
A plot of corrected ICL values for the models included in \code{x} (the list
of \code{coseq} objects)
}
\description{
Compare the corrected ICL values after applying the arcsin, logit, and logMedianRef
transformations in a coseq analysis
}
\examples{
## Simulate toy data, n = 300 observations
set.seed(12345)
countmat <- matrix(runif(300*4, min=0, max=500), nrow=300, ncol=4)
countmat <- countmat[which(rowSums(countmat) > 0),]
conds <- rep(c("A","B","C","D"), each=2)

## Run the Normal mixture model for K = 2,3,4
run_arcsin <- coseq(y=countmat, K=2:4, iter=5, transformation="arcsin")

## Plot and summarize results
plot(run_arcsin)
summary(run_arcsin)

## Compare ARI values for all models (no plot generated here)
ARI <- compareARI(run_arcsin, plot=FALSE)

## Compare ICL values for models with arcsin and logit transformations
run_logit <- coseq(y=countmat, K=2:4, iter=5, transformation="logit")
compareICL(list(run_arcsin, run_logit))

}
\author{
Andrea Rau, Cathy Maugis-Rabusseau
}

