% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coseq-package.R
\docType{package}
\name{coseq-package}
\alias{coseq-package}
\title{Co-expression and co-abundance analysis of high-throughput sequencing data}
\description{
Mixture models are implemented to cluster genes from high-throughput
transcriptome sequencing (RNA-seq) data. Parameter estimation is performed
using the EM algorithm, and model selection is performed using
either the slope heuristics or the integrated completed likelihood (ICL)
criterion.
}
\details{
\tabular{ll}{ Package: \tab coseq\cr Type: \tab Package\cr Version:
\tab 0.1.10\cr Date: \tab 2016-07-11\cr License: \tab GPL (>=3)\cr LazyLoad:
\tab yes\cr }
}
\examples{
## Simulate toy data, n = 300 observations
countmat <- matrix(rnorm(300*8), nrow=300, ncol=8)
conds <- rep(c("A","B","C","D"), each=2)

## Run the Normal mixture model for K = 2,3,4
run <- coseq(y=countmat, K=3:4, norm="none")
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau

Maintainer: Andrea Rau <\url{andrea.rau@jouy.inra.fr}>
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011)
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\keyword{cluster}
\keyword{models}

