% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coseq.R
\name{coseq}
\alias{coseq}
\title{Co-expression and co-abudance analysis of high-throughput sequencing data}
\usage{
coseq(y, K, conds = NULL, norm = "TMM", model = "Normal",
  transformation = "none", subset.index = NA, meanFilterCutoff = NULL,
  geneIDs = 1:nrow(y), parallel = TRUE, BPPARAM = bpparam(), ...)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value). If \code{fixed.lambda}
contains a list of lambda values to be fixed, \code{K} corresponds to the
number of clusters in addition to those fixed.}

\item{conds}{Vector of length \emph{q} defining the condition (treatment
group) for each variable (column) in \code{y}}

\item{norm}{The type of estimator to be used to normalize for differences in
library size: (\dQuote{\code{TC}} for total count, \dQuote{\code{UQ}} for
upper quantile, \dQuote{\code{Med}} for median, \dQuote{\code{DESeq}} for
the normalization method in the DESeq package, and \dQuote{\code{TMM}} for
the TMM normalization method (Robinson and Oshlack, 2010). Can also be a
vector (of length \emph{q}) containing pre-estimated library size estimates
for each sample.}

\item{model}{Type of mixture model to use (\dQuote{\code{Poisson}} or \dQuote{\code{Normal}})}

\item{transformation}{Transformation type to be used: \dQuote{\code{voom}}, \dQuote{\code{logRPKM}}
(if \code{geneLength} is provided by user), \dQuote{\code{arcsin}}, \dQuote{\code{logit}},
\dQuote{\code{logMedianRef}}, \dQuote{\code{profile}}, \dQuote{\code{none}}}

\item{subset.index}{Optional vector providing the indices of a subset of
genes that should be used for the co-expression analysis (i.e., row indices
of the data matrix \code{y}.}

\item{meanFilterCutoff}{Value used to filter low mean normalized counts if desired.}

\item{geneIDs}{Optional vector of gene ID's.}

\item{parallel}{If \code{FALSE}, no parallelization. If \code{TRUE}, parallel 
execution using BiocParallel (see next argument \code{BPPARAM}). A note on running 
in parallel using BiocParallel: it may be advantageous to remove large, unneeded objects 
from the current R environment before calling the function, as it is possible that R's 
internal garbage collection will copy these files while running on worker nodes.}

\item{BPPARAM}{Optional parameter object passed internally to \code{bplapply} when 
\code{parallel=TRUE}. If not specified, the parameters last registered with \code{register} 
will be used.}

\item{...}{Additional optional parameters.}
}
\value{
Object of class \code{coseq} containing the full \code{results} (itself an object
of class \code{NormMixClus} or \code{PoisMixClusWrapper}, the latter being the class
defined by the HTSCluster package), the \code{model} used, and if
applicable, the data \code{transformation} used.
}
\description{
Co-expression and co-abudance analysis of high-throughput sequencing data
}

