% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coseq-methods.R
\name{coseq}
\alias{coseq}
\alias{coseq.DESeqDataSet}
\alias{coseq.data.frame}
\alias{coseq.matrix}
\title{Co-expression or co-abudance analysis of high-throughput sequencing data}
\usage{
coseq(y, K, subset = NULL, model = "Normal", transformation = "none",
  norm = "TMM", meanFilterCutoff = NULL, modelChoice = "ICL",
  parallel = FALSE, BPPARAM = bpparam(), ...)

\method{coseq}{matrix}(y, K, subset = NULL, model = "Normal",
  transformation = "none", norm = "TMM", meanFilterCutoff = NULL,
  modelChoice = "ICL", parallel = FALSE, BPPARAM = bpparam(), ...)

\method{coseq}{data.frame}(y, K, subset = NULL, model = "Normal",
  transformation = "arcsin", norm = "TMM", meanFilterCutoff = NULL,
  modelChoice = "ICL", parallel = FALSE, BPPARAM = bpparam(), ...)

\method{coseq}{DESeqDataSet}(y, K, subset = NULL, model = "Normal",
  transformation = "arcsin", norm = "TMM", meanFilterCutoff = NULL,
  modelChoice = "ICL", parallel = FALSE, BPPARAM = bpparam(), ...)
}
\arguments{
\item{y}{(\emph{n} x \emph{q}) matrix of observed counts for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value or a vector of values)}

\item{subset}{Optional vector providing the indices of a subset of
genes that should be used for the co-expression analysis (i.e., row indices
of the data matrix \code{y}. For the generic function \code{coseq}, the results of a previously
run differential analysis may be used to select a subset of genes on which to perform the 
co-expression analysis. If this is desired, \code{subset.index} can also be an object of class 
DESeqResults (from the \code{results} function in \code{DESeq2}).}

\item{model}{Type of mixture model to use (\dQuote{\code{Poisson}} or \dQuote{\code{Normal}})}

\item{transformation}{Transformation type to be used: \dQuote{\code{voom}}, \dQuote{\code{logRPKM}}
(if \code{geneLength} is provided by user), \dQuote{\code{arcsin}}, \dQuote{\code{logit}},
\dQuote{\code{logMedianRef}}, \dQuote{\code{profile}}, \dQuote{\code{none}}}

\item{norm}{The type of estimator to be used to normalize for differences in
library size: (\dQuote{\code{TC}} for total count, \dQuote{\code{UQ}} for
upper quantile, \dQuote{\code{Med}} for median, \dQuote{\code{DESeq}} for
the normalization method in the DESeq package, and \dQuote{\code{TMM}} for
the TMM normalization method (Robinson and Oshlack, 2010). Can also be a
vector (of length \emph{q}) containing pre-estimated library size estimates
for each sample.}

\item{meanFilterCutoff}{Value used to filter low mean normalized counts if desired (by default,
set to a value of 50)}

\item{modelChoice}{Criterion used to select the best model. For Gaussian mixture models, 
\dQuote{\code{ICL}} (integrated completed likelihood criterion) is currently supported. For Poisson
mixture models, \dQuote{\code{ICL}}, \dQuote{\code{BIC}} (Bayesian information criterion), and a
non-asymptotic criterion calibrated via the slope heuristics  using either the \dQuote{\code{DDSE}}
(data-driven slope estimation) or \dQuote{\code{Djump}} (dimension jump) approaches may be used.
See the \code{HTSCluster} package documentation for more details about the slope heuristics approaches.}

\item{parallel}{If \code{FALSE}, no parallelization. If \code{TRUE}, parallel 
execution using BiocParallel (see next argument \code{BPPARAM}). A note on running 
in parallel using BiocParallel: it may be advantageous to remove large, unneeded objects 
from the current R environment before calling the function, as it is possible that R's 
internal garbage collection will copy these files while running on worker nodes.}

\item{BPPARAM}{Optional parameter object passed internally to \code{bplapply} when 
\code{parallel=TRUE}. If not specified, the parameters last registered with \code{register} 
will be used.}

\item{...}{Additional optional parameters.}
}
\value{
An S3 object of class \code{coseq} containing the following:
\item{results }{Object of class \code{NormMixClus} or 
\code{PoisMixClusWrapper}, the latter being the class defined by the 
HTSCluster package}
\item{model }{Model used, either \code{Normal} or \code{Poisson}}
\item{transformation }{Transformation used on the data}
\item{tcounts }{Transformed data using to estimate model}
\item{y_profiles }{Normalized profiles for use in plotting}
}
\description{
This is the primary user interface for the \code{coseq} package.
Generic S3 functions are implemented to perform co-expression or co-abudance analysis of 
high-throughput sequencing data, with or without data transformation, using mixture models. 
The supported classes are \code{matrix}, \code{data.frame}, \code{DESeqDataSet}, 
\code{DGEList}, \code{DGEExact}, \code{DGEGLM}, and \code{DGELRT}. The output of \code{coseq} 
is an S3 object of class \code{coseq}.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Perform coseq analysis for \code{matrix} class

\item \code{data.frame}: Perform coseq analysis for \code{data.frame} class

\item \code{DESeqDataSet}: Perform coseq analysis for \code{DESeqDataSet} class from \code{DESeq2} 
package
}}
\examples{
## Simulate toy data, n = 300 observations
set.seed(12345)
countmat <- matrix(runif(300*4, min=0, max=500), nrow=300, ncol=4)
countmat <- countmat[which(rowSums(countmat) > 0),]
conds <- rep(c("A","B","C","D"), each=2)

## Run the Normal mixture model for K = 2,3,4
run_arcsin <- coseq(y=countmat, K=2:4, iter=5, transformation="arcsin")

## Plot and summarize results
plot(run_arcsin)
summary(run_arcsin)

## Compare ARI values for all models (no plot generated here)
ARI <- compareARI(run_arcsin, plot=FALSE)

## Compare ICL values for models with arcsin and logit transformations
run_logit <- coseq(y=countmat, K=2:4, iter=5, transformation="logit")
compareICL(list(run_arcsin, run_logit))

}
\author{
Andrea Rau
}

