% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{logclr}
\alias{logclr}
\title{Calculate the Log Centered Log Ratio (logCLR) transformation}
\usage{
logclr(profiles)
}
\arguments{
\item{profiles}{Matrix of profiles. Note that the presence of 0 values causes an error message to be produced.}
}
\value{
logCLR-transformed profiles
}
\description{
Calculate the Log Centered Log Ratio (logCLR) transformation
}

