% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormMixClus.R
\name{NormMixClus}
\alias{NormMixClus}
\title{Normal mixture model estimation and selection for a series of cluster numbers}
\usage{
NormMixClus(y_profiles, K, subset.index = NA, parallel = TRUE,
  BPPARAM = bpparam(), ...)
}
\arguments{
\item{y_profiles}{(\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{K}{Number of clusters (a single value or a sequence of values).}

\item{subset.index}{Optional vector providing the indices of a subset of
genes that should be used for the co-expression analysis (i.e., row indices
of the data matrix \code{y}.}

\item{parallel}{If \code{FALSE}, no parallelization. If \code{TRUE}, parallel 
execution using BiocParallel (see next argument \code{BPPARAM}). A note on running 
in parallel using BiocParallel: it may be advantageous to remove large, unneeded objects 
from the current R environment before calling the function, as it is possible that R's 
internal garbage collection will copy these files while running on worker nodes.}

\item{BPPARAM}{Optional parameter object passed internally to \code{bplapply} when 
\code{parallel=TRUE}. If not specified, the parameters last registered with \code{register} 
will be used.}

\item{...}{Additional optional parameters to be passed to \code{\link{NormMixClus_K}}.}
}
\value{
\item{nbCluster.all }{Vector giving the number of clusters for each of the fitted models}
\item{loglike.all }{Log likelihoods calculated for each of the fitted models}
\item{ICL.all}{ICL values calculated for each of the fitted models}
\item{ICL.results }{Object of class
\code{NormMixClus} giving the results from the model chosen via the ICL
criterion} 
\item{all.results }{List of objects of class \code{NormMixClus} giving the results for all models}
}
\description{
Output of \code{NormMixClus} is an S3 object of class \code{NormMixClus}.
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau
}

