% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormMixParam.R
\name{NormMixParam}
\alias{NormMixParam}
\title{Calculate the mean and variance parameters for a normal mixture model}
\usage{
NormMixParam(x, y_profiles = NULL, modelChoice = NULL)
}
\arguments{
\item{x}{Object of class \code{coseq}, \code{NormMixClus}, or \code{NormMixClus_K}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables, required for \code{x} of class \code{NormMixClus}
or \code{NormMixClus_K}}

\item{modelChoice}{The model used for parameter estimation for objects \code{x} of
class \code{coseq} or \code{NormMixClus}. When \code{NULL}, the model selected
by the ICL criterion is used; otherwise, \code{modelChoice} should designate the number
of clusters in the desired model}
}
\value{
\item{pi }{ Vector of dimension \emph{K} with the estimated cluster proportions from
the Gaussian mixture model, where \emph{K} is the number of clusters}
\item{mu }{ Matrix of dimension \emph{K} x \emph{d} containing the estimated mean
vector from the Gaussian mixture model, where \emph{d} is the
number of samples in the data \code{y_profiles} and \emph{K} is the number of clusters}
\item{Sigma }{ Array of dimension \emph{d} x \emph{d} x \emph{K} containing the
estimated covariance matrices from the Gaussian mixture model, where \emph{d} is the
number of samples in the data \code{y_profiles} and \emph{K} is the number of clusters}
}
\description{
Corresponds to pK_Lk_Ck model
}

