% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coseq.R
\name{summary.coseq}
\alias{summary.coseq}
\title{Summarize results from clustering using a Poisson or Gaussian mixture model}
\usage{
\method{summary}{coseq}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"coseq"}}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Poisson or Gaussian
mixture model.
}
\examples{

set.seed(12345)

## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 2000 observations


}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{coseq}}
}
\keyword{methods}

