% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NormMixClus.R
\name{summary.NormMixClus}
\alias{summary.NormMixClus}
\title{Summarize results from clustering using a Normal mixture model}
\usage{
\method{summary}{NormMixClus}(object, y_profiles, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{"NormClus"}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{digits}{Integer indicating the number of decimal places to be used
for mixture model parameters}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Normal
mixture model estimated using \code{NormMixClus}.
}
\details{
The summary function for an object of class \code{"NormMixClus"}
provides the number of clusters selected for the ICL
model selection approach.
}
\examples{

set.seed(12345)

## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 2000 observations


}
\author{
Andrea Rau
}
\references{
Rau, A., Maugis-Rabusseau, C., Martin-Magniette, M.-L., Celeux,
G. (2015) Co-expression analysis of high-throughput transcriptome sequencing
data with Poisson mixture models. Bioinformatics, doi:
10.1093/bioinformatics/btu845.

Rau, A., Celeux, G., Martin-Magniette, M.-L., Maugis-Rabusseau, C. (2011).
Clustering high-throughput sequencing data with Poisson mixture models.
Inria Research Report 7786. Available at
\url{http://hal.inria.fr/inria-00638082}.
}
\seealso{
\code{\link{NormMixClus}}, \code{\link{NormMixClus_K}}
}
\keyword{methods}

