% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NormMixClus_K.R
\name{summary.NormMixClus_K}
\alias{summary.NormMixClus_K}
\title{Summarize results from clustering using a Normal mixture model}
\usage{
\method{summary}{NormMixClus_K}(object, y_profiles, digits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{"NormMixClus"}}

\item{y_profiles}{y (\emph{n} x \emph{q}) matrix of observed profiles for \emph{n}
observations and \emph{q} variables}

\item{digits}{Integer indicating the number of decimal places to be used
for mixture model parameters}

\item{...}{Additional arguments}
}
\description{
A function to summarize the clustering results obtained from a Normal
mixture model.
}
\details{
The summary function for an object of class \code{"NormMixClus_K"} provides the
following summary of results:

1) Number of clusters and model selection criterion used, if applicable.

2) Number of observations across all clusters with a maximum conditional
probability greater than 90% (and corresponding percentage of total
observations) for the selected model.

3) Number of observations per cluster with a maximum conditional probability
greater than 90% (and corresponding percentage of total observations per
cluster) for the selected model.

4) \eqn{\ensuremath\boldsymbol{\mu}}{\mu} values for the selected
model.

5) \eqn{\ensuremath\boldsymbol{\pi}}{\pi} values for the selected model.
}
\examples{
## Simulate data as shown in Rau et al. (2011)
## Library size setting "A", high cluster separation
## n = 200 observations

y <- matrix(rnorm(300*8), nrow=300, ncol=8)
conds <- rep(c("A","B","C","D"), each=2)

## Run the Normal mixture model for K = 3 and 4
run <- coseq(y, K = 3:4, norm="none")
}
\author{
Andrea Rau
}
\seealso{
\code{\link{NormMixClus}}, \code{\link{NormMixClus_K}}
}
\keyword{methods}

